pyproject.toml
[project]
name = "rlms-mlr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "gdown>=5.2.0",
    "gunicorn>=23.0.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "mlflow>=2.22.0",
    "optuna>=2.10.1",
    "optuna-dashboard==0.4",
    "optuna-fast-fanova>=0.0.4",
    "platformdirs>=4.3.7",
    "psycopg2>=2.9.10",
    "pytest>=8.3.5",
    "rich>=14.0.0",
    "sphinx>=8.1.3",
    "sqlalchemy==1.4.46",
    "tensorboard>=2.19.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "wandb>=0.19.10",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu126--"

[tool.setuptools.packages.find]
where = ["."]
include = ["rlms_mlr*"]
